Initial fraud proof block generation
This change provides functionality for watch tower to construct a
special block for the fraud proof and corresponding functionality in
validator to pick it up from Avail.

This change also modifies sequencer to inject invalid block into Avail
every now and then. It's not written to blockchain and therefore it
doesn't break the e2e functionality, but it does trigger the watch tower
fraud proof functionality for testing purposes.