Introduce `ActiveSequencers` interface for `Verifier`
`Verifier` implementation needs to know if claimed `Miner`, present in
block headers, belongs to active [staked] sequencers. To decouple block
verification implementation from blockchain instance and staking
contract, this change introduces an interface to query
`ActiveSequencers`. It also allows for easy mocking in unit tests to
control the verification logic as needed.

This is needed for successfull testing of new refactored watchtower
implementation.