Initial version of caching active sequencers decorator
This is a first version of caching active sequencers decorator
implementation. This is intented to be used by wrapping any
`staking.ActiveSequencers` implementation with this one and then using
the returned instance instead.

This is currently not used nor has any accompanied due to tricky nature
of testing this kind of implementation (requires blockchain events
etc.). This can be wired-in to future integration tests however so we'll
get there eventually.