Refactor Avail streaming client implementation
First iteration of the streaming client tried to handle historical reads
from defined offset to HEAD and watch of the HEAD in the same loop,
which introduced issues with timing, synchronization of different client
instances and resource consumption.

This revised implementation separates the reading of history from
defined offset and the watch stream of new HEADs. This should be more
resource economic and allow distinct instances to be better aligned with
each others on new blocks (of course those are never 100% in-sync).