Extract Avail dependencies from consensus impl (#80)
Construct & configure Avail dependencies outside of the consensus
implementation. This allows better testability since now the account
management, client and sender implementations can be fully controlled by
external code.

Example of this is to implement a wrapper for `avail.Sender`, which
allows interception of sent Avail blocks, in order to modify them to be
"malicious", and that way implement new tests cases in integration
tests.