Make Avail client more test friendly (#83)
This change extracts `avail.BlockStream` into an interface and changes
`avail.Client` to be factory for the block stream. This allows unit and
integration tests to implement their own mocked `avail.Client` which
returns controlled mock implementation of the `avail.BlockStream` as
well.

This then enables testability for logic that works with avail blocks,
such as sequencer and watchtower.