Add initial implementation for `snapshot` package
`snapshot` package provides functionality to wrap EVM state and
blockchain storages with snapshotting ability. These snapshots can be
sent to other peers which and then apply them directly to storages.
This way, an optimistic protocol can synchronize nodes across P2P
network without need for executing transactions, but with ability to
conform to consistency and integrity of the blockchain state.