Streamline `avail/consensus` construction (#95)
Now that we have our own `server` package in-place, we can remove the
last bits of references into pluggable/configurable consensus
construction. In this project there will always be only one hard-wired
consensus implementation.

With this change we can now also drop the workarounds that were needed
to wire in additional dependencies and configuration params into our
consensus implementation.

This change is a preparation to allow easy wiring of snapshotter into
consensus.