Delay sequencer activation on startup
When the sequencer joins the blockchain / node network, it must wait for
the first fresh Avail block window to begin. It can happen, that when
the sequencer gets staking effective to protocol smart contract, it
becomes the "current active sequencer" (i.e. the leader) in the middle
of the block production window. This would break the blockchain
consistency, because there's already an elected node performing the
active sequencer duty and producing blocks. To avoid this conflict, a
sequencer node must initially wait until the staking becomes visible in
the blockchain and when the first new Avail block window begins after
the stake been visible in the blockchain.