Perform staking in the main work goroutine (#125)
There was a deadlock with TxPool.AddTx() in staking, which was caused by
wrong ordering of actions.

The `consensus.Start()` is called in `server.NewServer()` _before_ the
`txpool.Start()` has been called. Because the TxPool hasn't been started
yet, the channel send inside the `txpool.AddTx()` will eventually block,
because the goroutine that consumes that channel hasn't been started
yet.

When moving the staking to actual work goroutine (that's driving the
sequencer / watchtower logic), the `AddTx()` will correctly block until
the txpool instance has been started. Effectively removing the deadlock.