avail.Sender.Send() without waiting for any status (#130)
In order to avoid delays in extrinsic submissions, `Send()` uses the
`SubmitExtrinsic()` from gsrpc client library, which doesn't wait for
any status report. This allows sequencer to batch multiple blocks on
Avail in quick fashion, because the current implementation of gsrpc
incurs relatively long delay when the response status is waited for.