Reduce initial Avail account deposit amount
Now that `accounts` tool actually deposits the requested amount of AVLs,
it takes a long time to complete all necessary transactions. Because
`avail-settlement` now has functionality to top-up the Avail account
balance when it runs low, the initial deposit doesn't need to be so big.
In order to speed up development: reduce initial amount of AVL deposit.