Embed Edge's `blockchain` under `pkg` w/ bugfix
Upstream Edge's `blockchain.Blockchain` has behavior that prevents us
from using their blockchain reorg (in our use case it tracks one of the
forks up to genesis block and panics). In order to win some time, it was
pulled and embedded here with the fix and couple related tests disabled.

This is needed in order to fix the fraud proof / dispute resolution
processing where the forking needs to happen correctly in all
situations.