Implement fraud server (#175)
Fraud server (HTTP interface) listens in given port and provides
`sync.Once` executor to perform fraud on-demand. By default the server
doesn't do anything, but when a request to `/fraud/prime` is made, the
`sync.Once` is re-newed and on next block generation the corresponding
sequencer will inject a malicious tx to generated block.